refactor to common DA v1 tests
