+++ test/minter-suite-minters/DA/MinterDAExp/MinterDAExp.common.ts
@@ -2,15 +2,15 @@ import { constants, expectRevert } from "@openzeppelin/test-helpers";
 import { expect } from "chai";
 import { ethers } from "hardhat";
 
-import { ONE_MINUTE, ONE_HOUR, ONE_DAY } from "../../util/constants";
+import { ONE_MINUTE, ONE_HOUR, ONE_DAY } from "../../../util/constants";
 
-import { Minter_Common } from "../Minter.common";
+import { Minter_Common } from "../../Minter.common";
 
 /**
- * These tests are intended to check common DALin functionality.
+ * These tests are intended to check common DAExp functionality.
  * @dev assumes common BeforeEach to populate accounts, constants, and setup
  */
-export const MinterDALin_Common = async () => {
+export const MinterDAExp_Common = async () => {
   describe("common minter tests", async () => {
     Minter_Common();
   });
@@ -28,24 +28,17 @@ export const MinterDALin_Common = async () => {
     });
 
     it("calculates the price correctly", async function () {
-      await ethers.provider.send("evm_mine", [
-        this.startTime + this.auctionStartTimeOffset,
-      ]);
+      for (let i = 1; i <= 5; i++) {
+        let ownerBalance = await this.accounts.user.getBalance();
+        let price = this.startingPrice;
+        for (let j = 0; j < i; j++) {
+          price = price.div(2);
+        }
 
-      const step = ONE_MINUTE * 8; // 480 seconds
-      const numSteps = 15;
-      for (let i = 1; i < numSteps; i++) {
-        let userBalance = await this.accounts.user.getBalance();
-        let a = ethers.BigNumber.from(i * step).mul(
-          this.startingPrice.sub(this.basePrice).toString()
-        );
-        let t = ethers.BigNumber.from(a.toString());
-        let price = this.startingPrice.sub(t.div(step * numSteps));
-        let contractPriceInfo = await this.minter
-          .connect(this.accounts.user)
-          .getPriceInfo(this.projectZero);
-        await ethers.provider.send("evm_mine", [
-          this.startTime + this.auctionStartTimeOffset + i * step,
+        await ethers.provider.send("evm_setNextBlockTimestamp", [
+          this.startTime +
+            this.auctionStartTimeOffset +
+            i * this.defaultHalfLife,
         ]);
         await this.minter
           .connect(this.accounts.user)
@@ -62,11 +55,10 @@ export const MinterDALin_Common = async () => {
           }),
           "Must send minimum value to mint!"
         );
-        let userDelta = (await this.accounts.user.getBalance()).sub(
-          userBalance
+        let ownerDelta = (await this.accounts.user.getBalance()).sub(
+          ownerBalance
         );
-        expect(userDelta.mul("-1").lte(contractPriceInfo.tokenPriceInWei)).to.be
-          .true;
+        expect(ownerDelta.mul("-1").lte(price)).to.be.true;
       }
     });
 
@@ -78,8 +70,8 @@ export const MinterDALin_Common = async () => {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           this.projectZero,
-          this.startTime + ONE_HOUR,
-          this.startTime + 2 * ONE_HOUR,
+          this.startTime + this.auctionStartTimeOffset,
+          this.defaultHalfLife,
           this.startingPrice,
           this.basePrice
         );
@@ -98,15 +90,13 @@ export const MinterDALin_Common = async () => {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           this.projectZero,
-          this.startTime + ONE_HOUR,
-          this.startTime + 2 * ONE_HOUR,
+          this.startTime + this.auctionStartTimeOffset,
+          this.defaultHalfLife,
           this.startingPrice,
           this.basePrice
         );
 
-      await ethers.provider.send("evm_mine", [
-        this.startTime + this.auctionStartTimeOffset + 2 * ONE_HOUR,
-      ]);
+      await ethers.provider.send("evm_mine", [this.startTime + 5 * ONE_HOUR]);
 
       let contractPriceInfo = await this.minter
         .connect(this.accounts.user)
@@ -117,14 +107,16 @@ export const MinterDALin_Common = async () => {
 
   describe("setAuctionDetails", async function () {
     it("cannot be modified mid-auction", async function () {
-      await ethers.provider.send("evm_mine", [this.startTime + ONE_HOUR]);
+      await ethers.provider.send("evm_mine", [
+        this.startTime + 2 * this.auctionStartTimeOffset,
+      ]);
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             this.projectZero,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + this.auctionStartTimeOffset,
+            this.defaultHalfLife,
             this.startingPrice,
             this.basePrice
           ),
@@ -140,8 +132,8 @@ export const MinterDALin_Common = async () => {
         .connect(this.accounts.artist)
         .setAuctionDetails(
           this.projectZero,
-          this.startTime + ONE_MINUTE,
-          this.startTime + 2 * ONE_HOUR,
+          this.startTime + this.auctionStartTimeOffset,
+          this.defaultHalfLife,
           this.startingPrice,
           this.basePrice
         );
@@ -156,8 +148,8 @@ export const MinterDALin_Common = async () => {
           .connect(this.accounts.additional)
           .setAuctionDetails(
             this.projectZero,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + this.auctionStartTimeOffset,
+            this.defaultHalfLife,
             this.startingPrice,
             this.basePrice
           ),
@@ -172,8 +164,8 @@ export const MinterDALin_Common = async () => {
           .connect(this.accounts.deployer)
           .setAuctionDetails(
             this.projectZero,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + this.auctionStartTimeOffset,
+            this.defaultHalfLife,
             this.startingPrice,
             this.basePrice
           ),
@@ -190,8 +182,8 @@ export const MinterDALin_Common = async () => {
           .connect(this.accounts.artist)
           .setAuctionDetails(
             this.projectZero,
-            this.startTime + ONE_MINUTE,
-            this.startTime + 2 * ONE_HOUR,
+            this.startTime + this.auctionStartTimeOffset,
+            this.defaultHalfLife,
             this.basePrice,
             this.startingPrice
           ),
@@ -257,63 +249,78 @@ export const MinterDALin_Common = async () => {
     });
   });
 
-  describe("enforce and broadcasts min auction length", async function () {
-    it("enforces min/max auction length constraint", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(this.projectZero);
-      // expect revert when creating a new project with min/max reversed
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            this.projectZero,
-            this.startTime + ONE_HOUR * 2,
-            this.startTime + ONE_HOUR,
-            this.startingPrice,
-            this.basePrice
-          ),
-        "Auction end must be greater than auction start"
-      );
-    });
-
-    it("enforces min auction length constraint", async function () {
+  describe("enforce and broadcasts auction half-life", async function () {
+    it("enforces half-life min/max constraint", async function () {
       await this.minter
         .connect(this.accounts.deployer)
         .resetAuctionDetails(this.projectZero);
       // expect revert when creating a new project with
-      const invalidLengthSeconds = 60;
+      const invalidHalfLifeSecondsMin = ONE_MINUTE;
       await expectRevert(
         this.minter
           .connect(this.accounts.artist)
           .setAuctionDetails(
             this.projectZero,
-            this.startTime + ONE_HOUR,
-            this.startTime + ONE_HOUR + invalidLengthSeconds,
+            this.startTime + this.auctionStartTimeOffset,
+            invalidHalfLifeSecondsMin,
             this.startingPrice,
             this.basePrice
           ),
-        "Auction length must be at least minimumAuctionLengthSeconds"
+        "Price decay half life must fall between min and max allowable values"
+      );
+
+      // expect revert when creating a new project with
+      const invalidHalfLifeSecondsMax = ONE_DAY;
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.artist)
+          .setAuctionDetails(
+            this.projectZero,
+            this.startTime + this.auctionStartTimeOffset,
+            invalidHalfLifeSecondsMax,
+            this.startingPrice,
+            this.basePrice
+          ),
+        "Price decay half life must fall between min and max allowable values"
       );
     });
 
-    it("emits event when min auction length is updated", async function () {
-      const newLengthSeconds = 3601;
-      // emits event when minimum auction length is updated
+    it("emits event when allowable half life range is updated", async function () {
+      const newMinSeconds = 60;
+      const newMaxSeconds = 6000;
+      // emits event when allowable half life range is updated
       await expect(
         this.minter
           .connect(this.accounts.deployer)
-          .setMinimumAuctionLengthSeconds(newLengthSeconds)
+          .setAllowablePriceDecayHalfLifeRangeSeconds(
+            newMinSeconds,
+            newMaxSeconds
+          )
       )
-        .to.emit(this.minter, "MinimumAuctionLengthSecondsUpdated")
-        .withArgs(newLengthSeconds);
+        .to.emit(this.minter, "AuctionHalfLifeRangeSecondsUpdated")
+        .withArgs(newMinSeconds, newMaxSeconds);
     });
 
-    it("validate setMinimumAuctionLengthSeconds ACL", async function () {
+    it("validate setAllowablePriceDecayHalfLifeRangeSeconds guards", async function () {
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.deployer)
+          .setAllowablePriceDecayHalfLifeRangeSeconds(600, 60),
+        "Maximum half life must be greater than minimum"
+      );
+      await expectRevert(
+        this.minter
+          .connect(this.accounts.deployer)
+          .setAllowablePriceDecayHalfLifeRangeSeconds(0, 600),
+        "Half life of zero not allowed"
+      );
+    });
+
+    it("validate setAllowablePriceDecayHalfLifeRangeSeconds ACL", async function () {
       await expectRevert(
         this.minter
           .connect(this.accounts.additional)
-          .setMinimumAuctionLengthSeconds(600),
+          .setAllowablePriceDecayHalfLifeRangeSeconds(60, 600),
         "Only Core whitelisted"
       );
     });
